renamed message-types to error-message-types and added ErrorCodes, removed unneeded ErrorTypes
