+++ src/@types/error-message-types.ts
@@ -1,11 +1,3 @@
-export enum SuccessMessageType {
-  SECRET_INITIALISED = 'SECRET_INITIALISED',
-  SUBMIT_SUCCESS = 'SUBMIT_SUCCESS',
-  INVOKE_3DSECURE_METHOD = 'INVOKE_3DSECURE_METHOD',
-  INVOKE_3DSECURE_CHALLENGE = 'INVOKE_3DSECURE_CHALLENGE',
-  CLOSE_3DSECURE_CHALLENGE = 'CLOSE_3DSECURE_CHALLENGE',
-}
-
 export enum ErrorMessageType {
   PAYSHEET_INIT_FAILED = 'PAYSHEET_INIT_FAILED',
   INVALID_PAY_SECRET = 'INVALID_PAY_SECRET',
@@ -15,13 +7,17 @@ export enum ErrorMessageType {
   INVALID_CARD_TYPE = 'INVALID_CARD_TYPE',
   INVALID_CARD_DETAILS = 'INVALID_CARD_DETAILS',
   SERVER_ERROR = 'SERVER_ERROR',
-  ALREADY_PROCESSED = 'ALREADY_PROCESSED',
   TIMEOUT = 'TIMEOUT',
-  INVALID_ACTION = 'INVALID_ACTION',
   NO_PAY_SECRET = 'NO_PAY_SECRET',
   UNKNOWN_ERROR = 'UNKNOWN_ERROR',
-  ENVIRONMENT_MISMATCH = 'ENVIRONMENT_MISMATCH',
   FAILED_TO_SUBMIT = 'FAILED_TO_SUBMIT',
+}
+
+export enum ErrorCodes {
+  PAY_REQUEST_INVALID_STATUS = 'PAY_REQUEST_INVALID_STATUS',
+  ENVIRONMENT_MISMATCH = 'ENVIRONMENT_MISMATCH',
+  PAY_REQUEST_SECRET_REQUIRED = 'PAY_REQUEST_SECRET_REQUIRED',
+  WALLET_INIT_FAILED = 'WALLET_INIT_FAILED',
 }
 
 export type TyroErrorMessage = { type: ErrorMessageType; message: string };
@@ -59,17 +55,9 @@ export const TyroErrorMessages: Record<ErrorMessageType, TyroErrorMessage> = {
     type: ErrorMessageType.SERVER_ERROR,
     message: 'Server error',
   },
-  [ErrorMessageType.ALREADY_PROCESSED]: {
-    type: ErrorMessageType.ALREADY_PROCESSED,
-    message: 'PayRequest already processed',
-  },
   [ErrorMessageType.TIMEOUT]: {
     type: ErrorMessageType.TIMEOUT,
     message: 'Timeout',
-  },
-  [ErrorMessageType.INVALID_ACTION]: {
-    type: ErrorMessageType.INVALID_ACTION,
-    message: 'Invalid action',
   },
   [ErrorMessageType.NO_PAY_SECRET]: {
     type: ErrorMessageType.NO_PAY_SECRET,
@@ -79,27 +67,11 @@ export const TyroErrorMessages: Record<ErrorMessageType, TyroErrorMessage> = {
     type: ErrorMessageType.UNKNOWN_ERROR,
     message: 'Unknown error occurred',
   },
-  [ErrorMessageType.ENVIRONMENT_MISMATCH]: {
-    type: ErrorMessageType.ENVIRONMENT_MISMATCH,
-    message: 'There is an environment mismatch. Check TyroProvider was initialised with the correct value for liveMode',
-  },
   [ErrorMessageType.FAILED_TO_SUBMIT]: {
     type: ErrorMessageType.FAILED_TO_SUBMIT,
     message: 'Failed to submit the payment',
   },
 };
-
-export interface SuccessMessage {
-  success: SuccessMessageType;
-}
-
-export interface Invoke3DSecureMethodMessage extends SuccessMessage {
-  threeDSMethodUrl: string;
-}
-
-export interface Invoke3DSecureChallengeMessage extends SuccessMessage {
-  challengeUrl: string;
-}
 
 export interface ErrorMessage {
   errorType: string;
@@ -107,5 +79,3 @@ export interface ErrorMessage {
   errorMessage: string;
   gatewayCode?: string;
 }
-
-export type MessageResponse = SuccessMessage | ErrorMessage;
